Feature/v2

+++ routes/api/tag-routes.js
@@ -63,18 +63,30 @@ router.put('/:id', async (req, res) => {
 });
 
 router.delete('/:id', async (req, res) => {
-  // delete on tag by its `id` value
   try {
-    let deleteTag = await Tag.destroy({
+    // get name of category before deleting
+    const tag = await Tag.findOne({
+      attributes: ['tag_name'],
       where: {
         id: req.params.id,
       }
     });
-    res.status(200).json({ message: 'Tag deleted successfully!'})
+    const tagName = tag.getDataValue('tag_name');
+    
+    // if categoryName is null, return 404, else delete
+    if (tagName === null) {
+      return res.status(404).json({ message: 'Oops, no tag found with this ID! Try another ID' });
+    } else {
+      let deleteTag = await Tag.destroy({
+        where: {
+          id: req.params.id,
+        },
+      });
+      res.status(200).json({ message: `${tagName} tag deleted successfully!` });
+    }
   }
   catch (err) {
-    res.status(404).json({ message: 'Oops, no tag found with this ID! Try another ID' })
-    res.status(500).json(err);
+    res.status(500).json(err); 
   }
 });
